+++ src/main/gen/nl/dirkgroot/structurizr/dsl/psi/impl/SDAnimationBlockImpl.java
@@ -10,33 +10,21 @@ import com.intellij.psi.util.PsiTreeUtil;
 import static nl.dirkgroot.structurizr.dsl.psi.SDTypes.*;
 import nl.dirkgroot.structurizr.dsl.psi.*;
 
-public class SDAnimationStatementImpl extends SDStatementImpl implements SDAnimationStatement {
+public class SDAnimationBlockImpl extends SDBlockImpl implements SDAnimationBlock {
 
-  public SDAnimationStatementImpl(@NotNull ASTNode node) {
+  public SDAnimationBlockImpl(@NotNull ASTNode node) {
     super(node);
   }
 
   @Override
   public void accept(@NotNull SDVisitor visitor) {
-    visitor.visitAnimationStatement(this);
+    visitor.visitAnimationBlock(this);
   }
 
   @Override
   public void accept(@NotNull PsiElementVisitor visitor) {
     if (visitor instanceof SDVisitor) accept((SDVisitor)visitor);
     else super.accept(visitor);
-  }
-
-  @Override
-  @NotNull
-  public SDAnimationBlock getAnimationBlock() {
-    return findNotNullChildByClass(SDAnimationBlock.class);
-  }
-
-  @Override
-  @NotNull
-  public SDAnimationKeyword getAnimationKeyword() {
-    return findNotNullChildByClass(SDAnimationKeyword.class);
   }
 
 }