+++ src/main/gen/nl/dirkgroot/structurizr/dsl/psi/impl/SDScriptBlockImpl.java
@@ -8,15 +8,15 @@ import com.intellij.psi.PsiElement;
 import com.intellij.psi.PsiElementVisitor;
 import com.intellij.psi.util.PsiTreeUtil;
 import static nl.dirkgroot.structurizr.dsl.psi.SDTypes.*;
-import com.intellij.extapi.psi.ASTWrapperPsiElement;
 import nl.dirkgroot.structurizr.dsl.psi.*;
 
-public class SDScriptBlockImpl extends ASTWrapperPsiElement implements SDScriptBlock {
+public class SDScriptBlockImpl extends SDBlockImpl implements SDScriptBlock {
 
   public SDScriptBlockImpl(@NotNull ASTNode node) {
     super(node);
   }
 
+  @Override
   public void accept(@NotNull SDVisitor visitor) {
     visitor.visitScriptBlock(this);
   }